+++ SampleProjects/NetworkLib/test/client.cpp
@@ -1,0 +1,7 @@
+#include <ArduinoUnitTests.h>
+// test for including <Client.h> without <Arduino.h>
+#include <Client.h>
+
+unittest(test) { assertTrue(true); }
+
+unittest_main()