+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Travis and Appveyor CI
 
 ### Fixed
+- Missing include of `IPAddress.h` in `Client.h`
 - Mismatches between library names in `library.properties` and the directory names, which can cause cryptic failures
 - `LibraryProperties` skips over parse errors instead of crashing: only lines with non-empty keys and non-nil values are recorded